TRQ-3142 Increase maximum time for $prologalarm
$prologalarm was ignored on the prologue for a job. Also when
the epilogue was run the $prologalarm value was
ignored if it was more the 300. Now the $prologalarm
value is always honored regardless of how large it is for both
prologue and epilogue scripts. The default timeout is still
300 seconds.

+++ src/resmom/test/mom_main/scaffolding.c
@@ -46,7 +46,7 @@ extern char *server_alias;
 const char *dis_emsg[10];
 long *log_event_mask = NULL;
 int rpp_dbprt = 0;
-extern unsigned int pe_alarm_time;
+extern long pe_alarm_time;
 struct connection svr_conn[PBS_NET_MAX_CONNECTIONS];
 struct config standard_config[2];
 struct config dependent_config[2];
